Improve robomongo Dockerfile to reduce image size
Run all the apt-get install on 1 Run command, then clear apt cache.
Also removed installing dbus and generating a new UUID, but instead use
an existing UUID in the image.
If this UUID is not there, when we start robomongo a warning message is
shown that file /etc/machine-id is not found.